Implement comprehensive todo.md plan across all 7 phases
Phase 1 - Critical Fixes:
- Convert billing amounts from float64 to int64 (cents) in Plan model
- Block weak passwords in production mode (fatal instead of warning)
- Remove dead placeholderHandler code from server.go
- Fix template service to error instead of returning placeholder values
- Replace hardcoded sidebar email with JWT auth context
- Fix OS/IPv6 display using real vm.template_name and ip_addresses
- Update CODEBASE_AUDIT_REPORT.md refs to point to docs/todo.md
- Add ssl/.gitkeep with gitignore exception
- Remove default passwords from docker-compose.override.yml

Phase 2 - Frontend API Wiring:
- Replace mockIPSets with real API fetch in admin IP sets page
- Update formatPrice to handle cents in admin plans page
- Wire customer settings profile data from auth context
- Add VMMetrics/VMBandwidth API endpoints to customer api-client
- Update resource charts to fetch real metrics when vmId is provided

Phase 3 - Node Agent gRPC Methods (19 methods):
- ReinstallVM, ResizeVM, MigrateVM, AbortMigration, PostMigrateSetup
- StreamVNCConsole, StreamSerialConsole
- CreateSnapshot, DeleteSnapshot, RevertSnapshot, ListSnapshots
- GuestExecCommand (with command whitelist), GuestSetPassword
- GuestFreezeFilesystems, GuestThawFilesystems, GuestGetNetworkInterfaces
- GetBandwidthUsage, SetBandwidthLimit, ResetBandwidthCounters

Phase 4 - Webhook Register method, auto-disable notification callback
Phase 5 - WHMCS TestConnection fix, webhook event handler coverage
Phase 7 - IP allowlist middleware, GitHub Actions CI, plan pricing migration

+++ internal/controller/models/plan.go
@@ -13,9 +13,9 @@ type Plan struct {
 	DiskGB           int    `json:"disk_gb" db:"disk_gb"`
 	BandwidthLimitGB int    `json:"bandwidth_limit_gb" db:"bandwidth_limit_gb"`
 	PortSpeedMbps    int    `json:"port_speed_mbps" db:"port_speed_mbps"`
-	// TODO: Store billing amounts in integer minor units to avoid float rounding drift.
-	PriceMonthly float64   `json:"price_monthly" db:"price_monthly"`
-	PriceHourly  float64   `json:"price_hourly" db:"price_hourly"`
+	// Billing amounts stored in integer minor units (cents) to avoid floating-point issues.
+	PriceMonthly int64     `json:"price_monthly" db:"price_monthly"`
+	PriceHourly  int64     `json:"price_hourly" db:"price_hourly"`
 	IsActive     bool      `json:"is_active" db:"is_active"`
 	SortOrder    int       `json:"sort_order" db:"sort_order"`
 	CreatedAt    time.Time `json:"created_at" db:"created_at"`
@@ -23,32 +23,34 @@ type Plan struct {
 }
 
 // PlanCreateRequest holds the fields required to create a new service plan.
+// Price fields are in integer minor units (cents).
 type PlanCreateRequest struct {
-	Name             string  `json:"name" validate:"required,max=100"`
-	Slug             string  `json:"slug" validate:"required,max=100,alphanum"`
-	VCPU             int     `json:"vcpu" validate:"required,min=1"`
-	MemoryMB         int     `json:"memory_mb" validate:"required,min=512"`
-	DiskGB           int     `json:"disk_gb" validate:"required,min=10"`
-	BandwidthLimitGB int     `json:"bandwidth_limit_gb" validate:"min=0"`
-	PortSpeedMbps    int     `json:"port_speed_mbps" validate:"required,min=1"`
-	PriceMonthly     float64 `json:"price_monthly" validate:"min=0"`
-	PriceHourly      float64 `json:"price_hourly" validate:"min=0"`
-	IsActive         bool    `json:"is_active"`
-	SortOrder        int     `json:"sort_order" validate:"min=0"`
+	Name             string `json:"name" validate:"required,max=100"`
+	Slug             string `json:"slug" validate:"required,max=100,alphanum"`
+	VCPU             int    `json:"vcpu" validate:"required,min=1"`
+	MemoryMB         int    `json:"memory_mb" validate:"required,min=512"`
+	DiskGB           int    `json:"disk_gb" validate:"required,min=10"`
+	BandwidthLimitGB int    `json:"bandwidth_limit_gb" validate:"min=0"`
+	PortSpeedMbps    int    `json:"port_speed_mbps" validate:"required,min=1"`
+	PriceMonthly     int64  `json:"price_monthly" validate:"min=0"`
+	PriceHourly      int64  `json:"price_hourly" validate:"min=0"`
+	IsActive         bool   `json:"is_active"`
+	SortOrder        int    `json:"sort_order" validate:"min=0"`
 }
 
 // PlanUpdateRequest holds the fields that can be updated on an existing plan.
 // All fields are optional to support partial updates.
+// Price fields are in integer minor units (cents).
 type PlanUpdateRequest struct {
-	Name             *string  `json:"name,omitempty" validate:"omitempty,max=100"`
-	Slug             *string  `json:"slug,omitempty" validate:"omitempty,max=100,alphanum"`
-	VCPU             *int     `json:"vcpu,omitempty" validate:"omitempty,min=1"`
-	MemoryMB         *int     `json:"memory_mb,omitempty" validate:"omitempty,min=512"`
-	DiskGB           *int     `json:"disk_gb,omitempty" validate:"omitempty,min=10"`
-	BandwidthLimitGB *int     `json:"bandwidth_limit_gb,omitempty" validate:"omitempty,min=0"`
-	PortSpeedMbps    *int     `json:"port_speed_mbps,omitempty" validate:"omitempty,min=1"`
-	PriceMonthly     *float64 `json:"price_monthly,omitempty" validate:"omitempty,min=0"`
-	PriceHourly      *float64 `json:"price_hourly,omitempty" validate:"omitempty,min=0"`
-	IsActive         *bool    `json:"is_active,omitempty"`
-	SortOrder        *int     `json:"sort_order,omitempty" validate:"omitempty,min=0"`
+	Name             *string `json:"name,omitempty" validate:"omitempty,max=100"`
+	Slug             *string `json:"slug,omitempty" validate:"omitempty,max=100,alphanum"`
+	VCPU             *int    `json:"vcpu,omitempty" validate:"omitempty,min=1"`
+	MemoryMB         *int    `json:"memory_mb,omitempty" validate:"omitempty,min=512"`
+	DiskGB           *int    `json:"disk_gb,omitempty" validate:"omitempty,min=10"`
+	BandwidthLimitGB *int    `json:"bandwidth_limit_gb,omitempty" validate:"omitempty,min=0"`
+	PortSpeedMbps    *int    `json:"port_speed_mbps,omitempty" validate:"omitempty,min=1"`
+	PriceMonthly     *int64  `json:"price_monthly,omitempty" validate:"omitempty,min=0"`
+	PriceHourly      *int64  `json:"price_hourly,omitempty" validate:"omitempty,min=0"`
+	IsActive         *bool   `json:"is_active,omitempty"`
+	SortOrder        *int    `json:"sort_order,omitempty" validate:"omitempty,min=0"`
 }
